Generalised @Check support from SwaggerRulesValidator to RulesValidator
