+++ server/src/main/java/de/zalando/zally/rule/SwaggerRulesValidator.kt
@@ -1,12 +1,10 @@
 package de.zalando.zally.rule
 
-import de.zalando.zally.rule.api.Check
 import de.zalando.zally.rule.zalando.InvalidApiSchemaRule
 import io.swagger.models.Swagger
 import io.swagger.parser.SwaggerParser
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
-import java.lang.reflect.Method
 
 /**
  * This validator validates a given Swagger definition based
@@ -30,25 +28,6 @@ class SwaggerRulesValidator(@Autowired rules: List<SwaggerRule>,
             ignores.map { it.toString() }
         } else {
             emptyList()
-        }
-    }
-
-    override fun validator(root: Swagger): (SwaggerRule) -> Iterable<Violation> {
-        return { rule: SwaggerRule ->
-            rule::class.java.methods
-                    .filter { it.isAnnotationPresent(Check::class.java) }
-                    .filter { it.parameters.size == 1 }
-                    .filter { it.parameters[0].type.isAssignableFrom(root::class.java) }
-                    .flatMap { invoke(it, rule, root) }
-        }
-    }
-
-    private fun invoke(check: Method, rule: SwaggerRule, root: Any): Iterable<Violation> {
-        val result = check.invoke(rule, root)
-        return when (result) {
-            null -> emptyList()
-            is Violation -> listOf(result)
-            else -> throw Exception("Unsupported return type for a @Check check!: ${result::class.java}")
         }
     }
 }
